+++ bundles/org.openhab.io.transport.modbus/src/main/java/org/openhab/io/transport/modbus/internal/ModbusManagerImpl.java
@@ -42,9 +42,6 @@ import org.openhab.io.transport.modbus.ModbusManagerListener;
 import org.openhab.io.transport.modbus.ModbusReadCallback;
 import org.openhab.io.transport.modbus.ModbusReadRequestBlueprint;
 import org.openhab.io.transport.modbus.ModbusRequestBlueprint;
-import org.openhab.io.transport.modbus.ModbusUnexpectedResponseFunctionCodeException;
-import org.openhab.io.transport.modbus.ModbusUnexpectedResponseSizeException;
-import org.openhab.io.transport.modbus.ModbusUnexpectedTransactionIdException;
 import org.openhab.io.transport.modbus.ModbusWriteCallback;
 import org.openhab.io.transport.modbus.ModbusWriteRequestBlueprint;
 import org.openhab.io.transport.modbus.PollTask;
@@ -56,6 +53,9 @@ import org.openhab.io.transport.modbus.endpoint.ModbusSlaveEndpoint;
 import org.openhab.io.transport.modbus.endpoint.ModbusSlaveEndpointVisitor;
 import org.openhab.io.transport.modbus.endpoint.ModbusTCPSlaveEndpoint;
 import org.openhab.io.transport.modbus.endpoint.ModbusUDPSlaveEndpoint;
+import org.openhab.io.transport.modbus.exception.ModbusUnexpectedResponseFunctionCodeException;
+import org.openhab.io.transport.modbus.exception.ModbusUnexpectedResponseSizeException;
+import org.openhab.io.transport.modbus.exception.ModbusUnexpectedTransactionIdException;
 import org.openhab.io.transport.modbus.internal.pooling.ModbusSlaveConnectionFactoryImpl;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
